✨ Add global properties file KRM functions
The directory values contains a `_properties.yaml` file containing the
most important properties of the platform. It also contains other files
composing a transformation pipeline:

- `_properties.yaml` is a generator for the properties.
- `argocd.yaml` contains the replacements for the argocd application.
- `zz_remove_local.yaml` removes the properties from the output and
  adds an annotation to all the resources to respect wide sequence
  indentation.

By running this command:

```console
> kustomize fn run --enable-exec --fn-path packages/argocd
```

The argocd application is customized according to the properties
contained in the properties file.

`_properties.yaml` starts with `_` so it comes first in the
transformation. `zz_remove_local.yaml` is prefixed with `zz_` so it
comes last.